llm-orchestration/requirements.txt
# Datclaw Memory Engine - LLM Orchestration Service
# Frozen from working Python 3.12 environment

# Web Framework
fastapi==0.109.0
uvicorn==0.27.0
starlette==0.35.1
pydantic==2.12.5
pydantic_core==2.41.5
pydantic-settings==2.1.0

# Redis
redis==5.0.1
hiredis==3.3.0

# ArangoDB
python-arango==7.9.0

# Qdrant
qdrant-client==1.7.0
grpcio==1.74.0
grpcio-tools==1.74.0

# LLM & Embeddings
openai==1.10.0
tiktoken==0.5.2
anthropic==0.18.0

# Vector & Text Processing
sentence-transformers==2.3.1
numpy==1.26.3
scipy==1.12.0

# NLP
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
nltk==3.9.2

# Time & Temporal Processing
python-dateutil==2.8.2
pytz==2024.1
tzdata==2025.2

# S3 Storage (Cold Archive)
boto3==1.34.34
botocore==1.34.34
s3transfer==0.10.4

# Async & Concurrency
aiohttp==3.9.3
aiosignal==1.4.0
motor==3.3.2

# Monitoring & Logging
prometheus-client==0.19.0
structlog==24.1.0
python-json-logger==2.0.7

# Configuration
PyYAML==6.0.1
python-dotenv==1.0.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-timeout==2.4.0
httpx==0.26.0

# Authentication
passlib==1.7.4
bcrypt==4.0.1
python-jose==3.5.0
cryptography==46.0.5

# Utilities
typing_extensions==4.15.0
tenacity==8.2.3
rich==13.7.0

# ML & Training
torch==2.5.1
transformers==4.57.1
scikit-learn==1.7.2
pandas==2.3.3
lightgbm==4.6.0
shap==0.49.1
numba==0.62.1
llvmlite==0.45.1
cloudpickle==3.1.2
slicer==0.0.8
matplotlib==3.10.7
joblib==1.5.2

# Spacy dependencies
thinc==8.3.10
blis==1.3.3
catalogue==2.0.10
confection==0.1.5
cymem==2.0.13
murmurhash==1.0.15
preshed==3.0.12
srsly==2.5.2
wasabi==1.1.3
weasel==0.4.3
smart_open==7.5.0

# Transformers/torch dependencies
tokenizers==0.22.1
safetensors==0.7.0
huggingface-hub==0.36.0
sentencepiece==0.2.1
regex==2025.11.3

# Common dependencies
anyio==4.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
filelock==3.20.0
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
idna==3.11
Jinja2==3.1.6
MarkupSafe==3.0.3
multidict==6.7.0
networkx==3.5
packaging==25.0
requests==2.32.5
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
urllib3==1.26.20
wrapt==2.0.1
yarl==1.22.0
